+++ time/time.go
@@ -1,2 +1,7 @@
 // Package time extends the time package in the stdlib.
 package time
+
+// Time format extensions for the ones in the standard library.
+const (
+	MySQLDateTime = "2006-01-02 15:04:05"
+)